+++ tests/Areas/AzureIsv/LiveTests/AzureIsvCommandTests.cs
@@ -23,7 +23,7 @@ public class AzureIsvCommandTests(LiveTestFixture liveTestFixture, ITestOutputHe
             Assert.Skip("Test skipped because Tenant is not 'Customer LED Tenant'.");
         }
         var result = await CallToolAsync(
-            "azmcp-datadog-monitoredresources-list",
+            "azmcp_datadog_monitoredresources_list",
             new()
             {
                 { "subscription", Settings.SubscriptionId },