+++ tests/Areas/Kusto/LiveTests/KustoCommandTests.cs
@@ -31,7 +31,7 @@ public class KustoCommandTests(LiveTestFixture liveTestFixture, ITestOutputHelpe
             var credentials = new DefaultAzureCredential();
             await Client.PingAsync();
             var clusterInfo = await CallToolAsync(
-                "azmcp-kusto-cluster-get",
+                "azmcp_kusto_cluster_get",
                 new()
                 {
                 { "subscription", Settings.SubscriptionId },
@@ -55,7 +55,7 @@ public class KustoCommandTests(LiveTestFixture liveTestFixture, ITestOutputHelpe
     public async Task Should_list_databases_in_cluster()
     {
         var result = await CallToolAsync(
-            "azmcp-kusto-database-list",
+            "azmcp_kusto_database_list",
             new()
             {
                 { "subscription", Settings.SubscriptionId },
@@ -72,7 +72,7 @@ public class KustoCommandTests(LiveTestFixture liveTestFixture, ITestOutputHelpe
     public async Task Should_list_kusto_tables()
     {
         var result = await CallToolAsync(
-            "azmcp-kusto-table-list",
+            "azmcp_kusto_table_list",
             new()
             {
                 { "subscription", Settings.SubscriptionId },
@@ -90,7 +90,7 @@ public class KustoCommandTests(LiveTestFixture liveTestFixture, ITestOutputHelpe
     public async Task Should_query_kusto()
     {
         var result = await CallToolAsync(
-            "azmcp-kusto-query",
+            "azmcp_kusto_query",
             new()
             {
                 { "subscription", Settings.SubscriptionId },